Move HLSParser to new `.hls` package
